requirements.txt
# 创建 requirements.txt
aiohttp==3.8.6
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
schedule==1.2.0
EOF
